Corrected some format specifiers/types (mstrings, otable)
git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2659 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/otable.c
@@ -280,7 +280,7 @@ show_otable_status (strbuf_t * sbuf, Bool verbose)
                    , obj_searches
                    , (float) obj_probes / (float) obj_searches);
         strbuf_addf(sbuf, "External lookups (succeed)   %"PRIuSTATCOUNTER" (%"PRIuSTATCOUNTER")\n"
-                   , (long)user_obj_lookups, (long)user_obj_found);
+                   , user_obj_lookups, user_obj_found);
 #if defined(__MWERKS__)
 #    pragma warn_largeargs reset
 #endif
